created a context to inherit the active games in lobby
