+++ api/battleship/room.py
@@ -2,9 +2,8 @@ from flask import (Blueprint, make_response, session, jsonify)
 
 from .auth import login_required
 from .extensions import socketio
+from .events import update_rooms, list_all_rooms
 from .helpers import generate_unique_code
-
-ROOMS = {}
 
 bp = Blueprint('room', __name__, url_prefix='/room')
 
@@ -12,10 +11,10 @@ bp = Blueprint('room', __name__, url_prefix='/room')
 @login_required
 def create_room():
   session['room'] = generate_unique_code()
-  ROOMS[(session['room'])] = {"gamestate": "someconfigs"}
-  socketio.emit('current_games', ROOMS)
+  latest_rooms = update_rooms(session['room'], {"gamestate": "someconfigs"}) 
+  socketio.emit('current_games', latest_rooms)
   return make_response({'room': session['room']}, 200)
 
 @bp.route('/list', methods=['GET'])
 def list_rooms():
-  return jsonify(ROOMS)
+  return make_response(list_all_rooms(), 200)
