+++ frontend/src/App.js
@@ -1,5 +1,6 @@
-import React from "react";
+import React, {useState, useEffect, createContext} from "react";
 import { useNavigate, Routes, Route } from "react-router-dom";
+import { socket } from "./socket";
 
 import Home from "./pages/Home/Home";
 import Game from "./pages/Game/Game";
@@ -7,8 +8,28 @@ import SignUpForm from "./pages/SignUp/SignUpForm";
 import LoginForm from "./pages/Login/LoginForm";
 import NotFound from "./pages/NotFound/NotFound";
 
+export const LobbyContext = createContext({})
+
 const App = () => {
+  
+  const [currentGames, setCurrentGames] = useState({});
+
+  function onCurrentGames(value) {
+    setCurrentGames((previous) => ({ ...previous, ...value }));
+  }
+
+  useEffect(() => {
+    socket.connect();
+    socket.on("current_games", onCurrentGames);
+
+    return () => {
+      socket.disconnect();
+      socket.off("current_games", onCurrentGames);
+    };
+  }, []);
+
   return (
+  <LobbyContext.Provider value={currentGames}>
     <Routes>
       <Route path="/" element={<Home navigate={useNavigate()} />} />
       <Route path="/signup" element={<SignUpForm navigate={useNavigate()} />} />
@@ -16,6 +37,7 @@ const App = () => {
       <Route path="/game/:gameId" element={<Game />} />
       <Route path="*" element={<NotFound />} />
     </Routes>
+  </LobbyContext.Provider>
   );
 };
